Implement endpoint to fetch all rectangular geofence coordinates.

+++ Fleet ManagementWebApplication/Models/RectangleGeofence.cs
@@ -4,9 +4,9 @@
     {
         public int ID { get; set; }
         public int GeofenceID { get; set; }
-        public int North { get; set; }
-        public int East { get; set; }
-        public int West { get; set; }
-        public int South { get; set; }
+        public float North { get; set; }
+        public float East { get; set; }
+        public float West { get; set; }
+        public float South { get; set; }
     }
 }
